Split JS into modules
